+++ kernel/thread/queue.c
@@ -13,7 +13,7 @@ queue_t ready_list;              /*  Struct with the ready_list root  */
 /* 'init_queues' sets all entries in the queue_table to initial values so they   *
  *  can be used safely later during OS operations.                               */
 void init_queues(void) {
-  for(int i = 0; i < NTHREADS; ++i) {
+  for(byte i = 0; i < NTHREADS; ++i) {
 	queue_table[i].key = -1;
 	queue_table[i].qnext = NULL;
 	queue_table[i].qprev = NULL;
@@ -84,7 +84,7 @@ int32_t dequeue_thread(queue_t* queue) {
 		return -1;
 	queue_t* node = queue->qnext;
 	
-	int threadid = 0;
+	byte threadid = 0;
 	for(; threadid < NTHREADS; ++threadid)
 		if(&queue_table[threadid] == node) break;
 	if(threadid == NTHREADS) return -1;
